+++ src/Terminals/SingleTerminal/SingleTerminalView.tsx
@@ -1,7 +1,9 @@
 import { IconButton } from '@mui/material';
 import { observer } from 'mobx-react-lite';
-import { useRef, ReactElement, useLayoutEffect } from 'react';
+import { useRef, ReactElement, useLayoutEffect, forwardRef, useEffect, useCallback, useMemo } from 'react';
 import ArrowDownwardIcon from '@mui/icons-material/ArrowDownward';
+import LockIcon from '@mui/icons-material/Lock';
+import LockOpenIcon from '@mui/icons-material/LockOpen';
 import { FixedSizeList } from 'react-window';
 
 import Terminal from './SingleTerminal';
@@ -21,8 +23,24 @@ interface RowProps {
   style: {};
 }
 
+// const outerListElement = forwardRef((props, ref: any) => (
+//   <div
+//     className="outerListElement"
+//     ref={ref}
+//     onWheel={(event) => {
+//       console.log('onWheel! event: ', event);
+//       if (event.deltaY < 0) {
+//         globalTerminal?.setScrollLock(false);
+//       }
+//     }}
+//     {...props} />
+// ))
+
+// let globalTerminal: Terminal | null = null;
+
 export default observer((props: Props) => {
   const { terminal, directionLabel, testId } = props;
+  // globalTerminal = terminal;
 
   const reactWindowRef = useRef<FixedSizeList>(null);
 
@@ -98,13 +116,16 @@ export default observer((props: Props) => {
   // is used, user sees jerky motion if scroll lock is applied
   // or they are in the middle of the data and data is being
   // removed from the start (buffer is full).
+  // This needs to be done because when we recreate the list it does not
+  // remember it's scroll position
   useLayoutEffect(() => {
     if (reactWindowRef.current === null) {
       return;
     }
     if (terminal.scrollLock) {
-      reactWindowRef.current.scrollToItem(terminal.filteredTerminalRows.length, 'auto');
-    } else {
+      reactWindowRef.current.scrollToItem(terminal.filteredTerminalRows.length - 1, 'auto');
+    } else
+    {
       // Scroll to the position determined by the Terminal model
       reactWindowRef.current.scrollTo(terminal.scrollPos);
     }
@@ -143,6 +164,40 @@ export default observer((props: Props) => {
       window.removeEventListener('resize', handleResize);
     };
   }, []);
+
+  // WARNING: Must use memoized component here, if not, it gets recreated on each render of
+  // the terminal and the scroll gets messed up. Spent a lot of time working this out :-
+  const outerListElementMemoized = useMemo(() => {
+    return forwardRef((props, ref: any) => (
+      <div
+        ref={ref}
+        onWheel={(event) => {
+          // Disable scroll lock if the user scrolled upwards
+          if (event.deltaY < 0) {
+            terminal.setScrollLock(false);
+          }
+        }}
+        {...props} />
+    ))
+  }, []);
+
+
+  let scrollLockUnlockIcon;
+  if (terminal.scrollLock) {
+    scrollLockUnlockIcon = <LockIcon
+      sx={{
+        width: '40px',
+        height: '40px',
+      }}
+    />
+  } else {
+    scrollLockUnlockIcon = <LockOpenIcon
+      sx={{
+        width: '40px',
+        height: '40px',
+      }}
+    />
+  }
 
   return (
     <>
@@ -228,28 +283,28 @@ export default observer((props: Props) => {
               terminal.fixedSizedListOnScroll(scrollProps);
             }}
             overscanCount={5}
+            outerElementType={outerListElementMemoized}
           >
             {Row}
           </FixedSizeList>
-          {/* ================== SCROLL LOCK ARROW ==================== */}
+          {/* ================== SCROLL LOCK/UNLOCK BUTTON ==================== */}
           <IconButton
             onClick={() => {
-              terminal.setScrollLock(true);
+              if (terminal.scrollLock) {
+                terminal.setScrollLock(false);
+              } else {
+                terminal.setScrollLock(true);
+              }
             }}
             sx={{
-              display: terminal.scrollLock ? 'none' : 'block',
+              // display: terminal.scrollLock ? 'none' : 'block',
               position: 'absolute', // Fix it to the bottom right of the TX/RX view port
-              bottom: '20px',
+              bottom: '10px',
               right: '30px',
               color: 'rgba(255, 255, 255, 0.4)',
             }}
           >
-            <ArrowDownwardIcon
-              sx={{
-                width: '40px',
-                height: '40px',
-              }}
-            />
+            {scrollLockUnlockIcon}
           </IconButton>
         </div>
       </div>